Proof of concept - close #13 close #14 (#15)
* disable sentry


* Add gov info


* log - different message for production


* Add examples template


* Change gov command to vote


* Add .gitignore file


* Add nodemon


* Add events handle template for button


* experiment with example


* Add eventsource


* Add SSE service


* Use data from SSE to create poll


* Use new EventSource endpoint


* WIP


* fix channel fetch, moved vote logic to dicord MessageReactionAdd event, added example code for axios call to get pol and to submit vote request

* Replace emoji with button


* Update button to contain poll id and option


* added create user+account if not exists. Voting works but reply still throws error

* vote counter

* added event listener for data request by front end - channel list

* add fetch role functionality

* fix interaction.reply() for ephemeral response but had to switch off slash commands

* added api-key auth, updated dependencies

* update changelog

* lint

* initialize slash-create and SSE event listeners similar to discordjs events, resturcture and refactor, add vote ephemeral user response

* update changelog

* update package info
