+++ src/app/app.ts
@@ -3,24 +3,21 @@ import Discord, { Client, ClientOptions, Intents, WSEventType } from 'discord.js
 import path from 'path';
 import fs from 'fs';
 import Log, { LogUtils } from './utils/Log';
-import constants from './service/constants/constants';
+import constants from './constants/constants';
 import NodeEventSource from 'eventsource';
 import axios from 'axios';
-const onSSEMessage = require('./service/sse');
 
-const client: Client = initializeClient();
-initializeEvents();
-
-const evtSource = new NodeEventSource(
-	constants.SSE_URL,
-	{ headers: { 'X-API-KEY': process.env.GOVERNATOR_API_KEY } },
-);
-onSSEMessage(evtSource, client);
-
+// set api key header for axios globally
 axios.defaults.headers.common = {
 	'X-API-KEY': process.env.GOVERNATOR_API_KEY,
 };
 
+// initialize discordjs client and events
+const client: Client = initializeClient();
+
+initializeDiscordJsEvents();
+
+// initialize slash-create
 const creator = new SlashCreator({
 	applicationID: process.env.DISCORD_BOT_APPLICATION_ID,
 	publicKey: process.env.DISCORD_BOT_PUBLIC_KEY,
@@ -28,28 +25,15 @@ const creator = new SlashCreator({
 	disableTimeouts: true,
 });
 
-creator.on('debug', (message) => Log.debug(`debug: ${ message }`));
-creator.on('warn', (message) => Log.warn(`warn: ${ message }`));
-creator.on('error', (error: Error) => Log.error(`error: ${ error }`));
-creator.on('synced', () => Log.debug('Commands synced!'));
-creator.on('commandRegister', (command: SlashCommand) => Log.debug(`Registered command ${command.commandName}`));
-creator.on('commandError', (command: SlashCommand, error: Error) => Log.error(`Command ${command.commandName}:`, {
-	indexMeta: true,
-	meta: {
-		name: error.name,
-		message: error.message,
-		stack: error.stack,
-		command,
-	},
-}));
+initializeSlashCreateEvents();
 
-// FIXME find a way to get slash commands to work without interfering with other INTERACTION_CREATE events
-// creator.on('componentInteraction', (interaction) => interaction.send({ content:'test' }));
+// initialize governator SSE
+const evtSource = new NodeEventSource(
+	constants.SSE_URL,
+	{ headers: { 'X-API-KEY': process.env.GOVERNATOR_API_KEY } },
+);
 
-// Ran after the command has completed
-creator.on('commandRun', (command:SlashCommand, result: Promise<any>, ctx: CommandContext) => {
-	LogUtils.logCommandEnd(ctx);
-});
+initializeGovernatorEvents();
 
 // Register command handlers
 creator
@@ -59,8 +43,7 @@ creator
 	.registerCommandsIn(path.join(__dirname, 'commands'))
 	.syncCommands();
 
-// Log client errors
-client.on('error', Log.error);
+
 // reload client
 client.destroy();
 
@@ -85,10 +68,13 @@ function initializeClient(): Client {
 	return new Discord.Client(clientOptions);
 }
 
-function initializeEvents(): void {
-	const eventFiles = fs.readdirSync(path.join(__dirname, '/events-discord')).filter(file => file.endsWith('.js'));
+function initializeDiscordJsEvents(): void {
+	// Log client errors
+	client.on('error', Log.error);
+
+	const eventFiles = fs.readdirSync(path.join(__dirname, '/events/events-discordjs')).filter(file => file.endsWith('.js'));
 	eventFiles.forEach(file => {
-		const event = new (require(`./events-discord/${file}`).default)();
+		const event = new (require(`./events/events-discordjs/${file}`).default)();
 		try {
 			if (event.once) {
 				client.once(event.name, (...args) => event.execute(...args, client));
@@ -108,5 +94,77 @@ function initializeEvents(): void {
 		}
 	});
 }
+
+function initializeSlashCreateEvents(): void {
+	creator.on('debug', (message) => Log.debug(`debug: ${ message }`));
+	creator.on('warn', (message) => Log.warn(`warn: ${ message }`));
+	creator.on('error', (error: Error) => Log.error(`error: ${ error }`));
+	creator.on('synced', () => Log.debug('Commands synced!'));
+	creator.on('commandRegister', (command: SlashCommand) => Log.debug(`Registered command ${command.commandName}`));
+	creator.on('commandError', (command: SlashCommand, error: Error) => Log.error(`Command ${command.commandName}:`, {
+		indexMeta: true,
+		meta: {
+			name: error.name,
+			message: error.message,
+			stack: error.stack,
+			command,
+		},
+	}));
+	// Ran after the command has completed
+	creator.on('commandRun', (command:SlashCommand, result: Promise<any>, ctx: CommandContext) => {
+		LogUtils.logCommandEnd(ctx);
+	});
+
+	const eventFiles = fs.readdirSync(path.join(__dirname, '/events/events-slash-create')).filter(file => file.endsWith('.js'));
+	eventFiles.forEach(file => {
+		const event = new (require(`./events/events-slash-create/${file}`).default)();
+		try {
+			if (event.once) {
+				creator.once(event.name, (...args) => event.execute(...args, client));
+			} else {
+				creator.on(event.name, (...args) => event.execute(...args, client));
+			}
+		} catch (e) {
+			Log.error('Slash-create event failed to process', {
+				indexMeta: true,
+				meta: {
+					name: e.name,
+					message: e.message,
+					stack: e.stack,
+					event,
+				},
+			});
+		}
+	});
+}
+
+function initializeGovernatorEvents(): void {
+	evtSource.onerror = function(err) {
+		if (err) {
+			if (err.status === 401 || err.status === 403) {
+				Log.error('not authorized');
+			}
+		}
+	};
+	const eventFiles = fs.readdirSync(path.join(__dirname, '/events/events-governator')).filter(file => file.endsWith('.js'));
+	eventFiles.forEach(file => {
+		const event = new (require(`./events/events-governator/${file}`).default)();
+		try {
+			evtSource.addEventListener(event.name, async (...args) => event.execute(...args, client));
+		} catch (e) {
+			console.log(e);
+			Log.error('Governator event failed to process', {
+				indexMeta: true,
+				meta: {
+					name: e.name,
+					message: e.message,
+					stack: e.stack,
+					event,
+				},
+			});
+		}
+	});
+}
+
 
 export default client;
